+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed 
+- Create ticket command supports message processing as optional.
+
 ### Added
 
 - Added duploctl command to create AI Helpdesk Ticket & Send AI Message.